🗑️ Demote pseudo-features to lib/ and their owning features
Apply the feature admission test to the five single-file pseudo-features:

- features/analytics/posthog.ts → lib/posthog.ts (cross-cutting infra)
- features/theme/client.tsx → lib/theme.tsx
- features/subscription/schema.ts → merged into features/billing/schema.ts
- features/google/service.ts → inlined into calendars/services/google-calendar.ts (its only consumer)
- features/developer/data.ts → features/api-keys/data.ts

navigation, feedback and quick-create stay: quick-create gained real
server logic, feedback owns the submit lifecycle (in frozen tRPC), and
navigation is domain-aware app chrome with no other home. Rename
navigation/config.tsx → client.tsx to fit the vocabulary.

Shrinks the vocabulary grandfather list to the two calendars service.ts
entries and removes four entries from the biome import allowlist.
