+++ apps/web/src/app/api/better-auth/[...all]/route.ts
@@ -1,6 +1,6 @@
 import { toNextJsHandler } from "better-auth/next-js";
-import { withPostHog } from "@/features/analytics/posthog";
 import { authLib } from "@/lib/auth";
+import { withPostHog } from "@/lib/posthog";
 
 const { POST: authPost, GET: authGet } = toNextJsHandler(authLib);
