+++ apps/web/src/features/api-keys/data.ts
@@ -1,8 +1,8 @@
 import "server-only";
 
-import { posthog } from "@/features/analytics/posthog";
 import type { SpaceTier } from "@/features/space/schema";
 import { isSelfHosted } from "@/lib/constants";
+import { posthog } from "@/lib/posthog";
 
 /**
  * Determines if a user has access to API features (API keys, developer tools)
