--- apps/web/src/features/subscription/schema.ts
@@ -1,17 +1,0 @@
-import * as z from "zod";
-
-export const subscriptionCheckoutMetadataSchema = z.object({
-  userId: z.string(),
-  spaceId: z.string().optional(),
-});
-
-export type SubscriptionCheckoutMetadata = z.infer<
-  typeof subscriptionCheckoutMetadataSchema
->;
-
-export const subscriptionMetadataSchema = z.object({
-  userId: z.string(),
-  spaceId: z.string(),
-});
-
-export type SubscriptionMetadata = z.infer<typeof subscriptionMetadataSchema>;
