+++ apps/web/src/trpc/routers/polls/comments.ts
@@ -6,9 +6,9 @@ import { TRPCError } from "@trpc/server";
 import { after } from "next/server";
 import * as z from "zod";
 import { getInstanceBranding } from "@/emails/branding";
-import { posthog } from "@/features/analytics/posthog";
 import { getNotificationRecipient } from "@/features/notifications/data";
 import { hasPollAdminAccess } from "@/features/poll/data";
+import { posthog } from "@/lib/posthog";
 import {
   createRateLimitMiddleware,
   publicProcedure,
