+++ apps/web/src/trpc/routers/user.ts
@@ -3,12 +3,12 @@ import { prisma } from "@rallly/database";
 import { sendRawEmail } from "@rallly/emails";
 import { TRPCError } from "@trpc/server";
 import * as z from "zod";
-import { posthog } from "@/features/analytics/posthog";
 import { feedbackSchema } from "@/features/feedback/schema";
 import { defaultNotificationPreferences } from "@/features/notifications/constants";
 import { getNotificationPreferences } from "@/features/notifications/data";
 import { activityEventTypes } from "@/features/notifications/schema";
 import { defineAbilityFor } from "@/features/user/ability";
+import { posthog } from "@/lib/posthog";
 import {
   createRateLimitMiddleware,
   privateProcedure,
